Add weather, qr_encode, and pastebin
These utilities all require an Internet connection to work. `weather` is
convenient for seeing where a Nerves device is (via GeoIP) and what the
weather is there. `qr_encode` is useful for copying chunks of data like
serial numbers to your phone, and `pastebin` is useful for sharing logs
and other text.