Add file path autocomplete
This works my looking at the beginning of a string. If it looks like
you're typing a file path and you press tab, it will try to complete it
or offer options.

Currently paths that start with `/` and `./` work. Nothing complicated
is supported, but it seems to be passible for what I've tried so far.