Add httpget/2
This utility covers a subset of curl like commands.
It performs an HTTP GET request on a specified URL and can print
or save the response to a local path.
It can display headers with a verbose switch.
It will auto add http:// if this is left off the URL.