This adds a unit test for the cat helper function.
It also changes the IO use write/1 instead of puts/1 due to an extra space produce by puts/1
this is due the fact that actual Unix cat command do not produce extra line.
And it  adds alittle information to the cat helper function definition
to indicate it prints the content of a file.