Add log_attach and log_detach
These are convenience functions for sending log messages to the current
IEx session. They reuse the Elixir console logger backend, but configure
it to report to the current group leader so that it works well with the
IEx prompt. Additionally, it monitors the group leader so that if it
goes away (like by terminating an ssh session), the logger backend will
be cleaned up.

The functionality here is similar to that provided by the `RingLogger`
library's `attach` and `detach` functions. These, however, are primarily
convenience functions for configuring the Elixir console logger backend.