Remove most module docs
We'd like to discourage users from calling functions outside of the
`Toolshed` module. However, we'd also like to have the documentation for
the helper functions exist with the implementation. We can do that for
`h/1`, but not with `ex_doc`. Since `ex_doc` will give lots of warnings
for `@moduledoc false`, just leave it blank. The real documentation
should be on the function anyway.