Merge helper modules into one large beam file
Toolshed helper functions have always been callable from the Toolshed
module. This has been done a couple ways, and the intent was any
function not in the Toolshed module was private. This commit doesn't
change this.

The problem this solves is that it's much easier to maintain Toolshed
helper functions when their code is in separate files. Each file is
smaller and it gets rid of debates over which file to best categorize a
function. However, in some environments, loading lots of BEAMs is
surprisingly slow. Additionally, this saves the repetition between BEAMs
so there's a minor space savings as well.

Less important benefits are:

1. Function help is only stored once and no hacks are needed to have it
   stored with the implementation module and accessible with the
   Toolshed module.
2. Helper selection is done at compile-time. If you're not running
   Nerves, no files with the word "Nerves" in them are even compiled by
   the Elixir compiler.
3. It's possible to do finegrain selection of helpers. This hasn't been
   implemented.